optimize solver to not repeat itself. works, but something seems off with optimization

+++ src/solver.h
@@ -8,7 +8,7 @@ void solve_corners(uint32_t* cube, uint32_t delay);
 
 /* This generates the order to solve edge pieces similarly
     to the human method to solve a cube blindfolded. */
-int solve_edges(uint32_t* cube, uint32_t delay);
+uint8_t solve_edges(uint32_t* cube, uint32_t delay);
 
 void solve(uint32_t* cube, uint32_t delay);
